feat(ui): split renderer layout metadata into modules
